More idiomatic protocol implementation w/extentions
